+++ lib/oli/delivery/sections/blueprint.ex
@@ -355,7 +355,9 @@ defmodule Oli.Delivery.Sections.Blueprint do
       on:
         section.id ==
           community_visibility.section_id and community_visibility.community_id == ^community_id,
-      where: is_nil(community_visibility.id) and section.type == :blueprint,
+      where:
+        is_nil(community_visibility.id) and section.type == :blueprint and
+          section.status == :active,
       select: section
     )
     |> Repo.all()